+++ Common/src/main/java/de/markusbordihn/easymodelentities/entity/EasyModelWaterHostEntity.java
@@ -20,6 +20,7 @@
 package de.markusbordihn.easymodelentities.entity;
 
 import de.markusbordihn.easymodelentities.api.data.EasyModelAnimationSetting;
+import de.markusbordihn.easymodelentities.api.data.EasyModelTextureSetting;
 import de.markusbordihn.easymodelentities.data.profile.ModelBodyType;
 import de.markusbordihn.easymodelentities.network.syncher.EasyModelEntityDataSerializers;
 import de.markusbordihn.easymodelentities.runtime.EasyModelRuntimeContract;
@@ -74,6 +75,9 @@ public abstract class EasyModelWaterHostEntity extends WaterAnimal implements Ea
       SynchedEntityData.defineId(EasyModelWaterHostEntity.class, EntityDataSerializers.BOOLEAN);
   private static final EntityDataAccessor<Boolean> RANDOM_STROLL =
       SynchedEntityData.defineId(EasyModelWaterHostEntity.class, EntityDataSerializers.BOOLEAN);
+  private static final EntityDataAccessor<EasyModelTextureSetting> TEXTURE =
+      SynchedEntityData.defineId(
+          EasyModelWaterHostEntity.class, EasyModelEntityDataSerializers.TEXTURE_SETTING);
 
   private static final EasyModelHostFields FIELDS =
       new EasyModelHostFields(
@@ -86,7 +90,8 @@ public abstract class EasyModelWaterHostEntity extends WaterAnimal implements Ea
           BODY_TYPE,
           ANIMATION_STATE,
           LOOK_AT_PLAYERS,
-          RANDOM_STROLL);
+          RANDOM_STROLL,
+          TEXTURE);
 
   private EasyModelRuntimeContract runtimeContract;
 
@@ -171,6 +176,16 @@ public abstract class EasyModelWaterHostEntity extends WaterAnimal implements Ea
   @Override
   public void setEasyModelAnimation(EasyModelAnimationSetting animation) {
     EasyModelHostSupport.setAnimation(this.entityData, FIELDS, animation);
+  }
+
+  @Override
+  public EasyModelTextureSetting getEasyModelTextureSetting() {
+    return EasyModelHostSupport.getTexture(this.entityData, FIELDS);
+  }
+
+  @Override
+  public void setEasyModelTexture(EasyModelTextureSetting texture) {
+    EasyModelHostSupport.setTexture(this.entityData, FIELDS, texture);
   }
 
   @Override